requirements.txt
pandas==2.0.2
requests==2.31.0
datarobot
datarobotx==0.1.20
streamlit
snowflake-connector-python
